split load_index
